SuperGCP/Cloud-Run-Functions/bigquery-to-gsheets/requirements.txt
gspread==5.8.0
gspread-dataframe==3.3.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
PyDrive==1.3.1
pandas
google-cloud-bigquery==3.9.0
google-cloud-storage==2.8.0
google-cloud-secret-manager==2.16.1
db-dtypes==1.1.1
functions-framework==3.*